COVID-19 Utility (Webpage) v1.2
The first version of the COVID-19 Utility (webpage). You can run it on your IDE, which will require you to copy and paste the streamlit command into your CMD, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the webpage. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is the second minor patch of the first release of the program.

Changes:
- Added more Python libraries for added informative visualizations
- Sidebar and 'selected' section have been updated
- Date and Time has been included
- Minor edits and tweaks, general improvement