COVID-19 Utility (Webpage) v1.3
COVID-19 Utility (Webpage) v1.3
The first version of the COVID-19 Utility (webpage). You can run it on your IDE, which will require you to copy and paste the streamlit command into your CMD, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the webpage. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is the third minor patch of the first release of the program.

Changes:
- Moved State Dictionary to separate python file
- Sidebar and 'selected' section have been updated to add more information and descriptive text
- Minor additions to codebase regarding sorting statistics and reading JSON/CSV files
- Cleaned up codebase
- Minor edits and tweaks, general improvements