COVID-19 Utility (Webpage) v1.3.3
COVID-19 Utility (Webpage) v1.3.3
The first version of the COVID-19 Utility (webpage). You can run it on your IDE, which will require you to copy and paste the streamlit command into your CMD, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the webpage. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a minor patch for the first release of the program.

Changes:
- Added accurate COVID data functionality for Coronavirus deaths from the COVIDActNow API, relating to the specified location
- Updated dictionary info for pulling more API data, now and for future updates
- Slightly tweaked Web Links
- Added hyperlinks to COVIDActNow API references in the data
- Nulled out error-prone functionality to be worked on later
- Cleaned up codebase
- Minor edits and tweaks, general improvements

Issues: (To be fixed)
- Program outputs accurate COVID death data from the COVIDActNow API, but needs fixing for accurate COVID case rate
- Webscraping functionality for national U.S COVID data does not work properly
- Missing Johns Hopkins COVID Data API functionality 
- Missing statistics and infographics relating to COVID data