COVID-19 Utility (Webpage) v1.3.4
COVID-19 Utility (Webpage) v1.3.4
The first version of the COVID-19 Utility (webpage). You can run it on your IDE, which will require you to copy and paste the streamlit command into your CMD, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the webpage. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a minor patch for the first release of the program.

Changes:
- Added foundations for matplotlib functionality to show data graphs in real time
- Slightly tweaked dictionary information
- Backend has been updated for future visualizations of data, in a nice manner
- Slight codebase tweaking
- Cleaned up codebase (slightly)
- Minor edits and tweaks, general improvements

Issues: (To be fixed)
- Program outputs accurate COVID death data from the COVIDActNow API, but needs fixing for accurate COVID case rate, not accurate yet
- Webscraping functionality for national U.S COVID data does not work properly
- Missing Johns Hopkins COVID Data API functionality 
- Missing statistics and infographics relating to COVID data 
- Console/Terminal prints out random number data before prompt is shown

Roadmap:
- Adding permissions for location access to be used for gathering COVID Data, as an alternative to selecting
- Updating selectbox to allow for state and county options to be used rather than typing in to console/terminal
- Making the webpage more visually interesting to look at
- Making the webpage into a hosted web application (via Streamlit Apps/GitHub pages)
- Adding email notification system
- Adding more COVID-19 Data statistics, such as ICU Bed Capacity and Community Levels (per CDC)