COVID-19 Utility (Web Version) v1.5(C)
COVID-19 Utility (Web Version) v1.5(C)

The first version of the COVID-19 Utility (Web Version). You can run it on your IDE, which will require you to copy and paste the streamlit command into your CMD, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the webpage. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a major revamp patch for the first release of the program.


**Developer note**
I have been working for months to configure API --> dict() ---> accurate COVID location data, and each patch sent out did not address the fundamental issue
that having more API sources brought.  I have essentially hit the development wall regarding this issue for a while. 

A few nights ago I stumbled across an organization that had created an API middleman for COVID API sources, with the dict() ---> accurate COVID location data.
This organization's middleman API fixes the API config --> accurate COVID location data issue, do I need to still work on the config?

Version 1.5(C) (C = Configured) brings a new version of the COVID-19-Utility (Web Version). This version has stripped out much of the old code for the older self-built configuration and instead swaps in the Disease.sh API middleman for easy access to the same API sources & data as before but with less errors! (and code)

I have not yet determined yet if I will continue to move forward with the (C) version with the Disease.SH API, but I wanted to push this out as a potential new way forward for this program, a different path to ultimately reaching the program's end goal. 

**Streamlit note**
Streamlit is not yet available with this version, the localized client version of v1.5(C) will work through your IDE but currently there is no Web Application version of 1.5(C), v.1.4.5 is the current version hosted, but this will be updated to reflect the changes soon.

Changes:
- Added new configured API middleman for NYT/JHU/CovidActNow/Worldometers API(s) & data
- Revamped codebase and removed most older dict() config code from (1.4.5 and older)
- Cleaned up codebase!

Issues: (To be fixed)
- Missing API sources
- NYT county data is having issues retrieving data
- Missing statistics and infographics relating to COVID data 

Roadmap:
- Adding rest of NYT/JHU/COVIDActNow API terms and functionality for more COVID data
- Adding permissions for location access to be used for gathering COVID Data, as an alternative to selecting
- Updating selectbox to allow for state and county options to be used rather than typing in to console/terminal
- Making the webpage more visually interesting to look at
- Making the webpage into a hosted web application (via Streamlit Apps/GitHub pages)
- Adding email notification system
- Adding more COVID-19 Data statistics, such as ICU Bed Capacity and Community Levels (per CDC)
- COVID-19 Data Predictive Model (To Be Determined)