COVID-19 Utility (Web Version) v1.6.1
COVID-19 Utility (Web Version) v1.6.1

The first version of the COVID-19 Utility (Web Version). You can run it on your IDE, which will require you to copy and paste the Streamlit command into your CMD/Terminal, opening the webpage onto your computer's browser. From there you will need to input the corresponding U.S State/County data into **the webpage** from there the COVID statistics will display. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a minor patch for the second (Pre)release of the program.


Changes:
- Added back Quick Web Links
- Made data output on Streamlit webpage more clean to look at and view
- Cleaned up Sidebar (for now)
- Cleaned up codebase

Issues: (To be fixed)
- CDC County data is having issues retrieving data
- Missing statistics and infographics relating to COVID data 

Roadmap:
- Adding rest of CDC API terms and functionality for more COVID data
- Adding permissions for location access to be used for gathering COVID Data, as an alternative to selecting
- Making the webpage more visually interesting to look at
- Making the webpage into a hosted web application (via Streamlit Apps/GitHub pages)
- Adding email notification system
- Adding TwitterBot functionality (To Be Determined)
- Adding more COVID-19 Data statistics, such as ICU Bed Capacity and Community Levels (per CDC)
- COVID-19 Data Predictive Model (To Be Determined)